feat: Refactores location of worker code to add more workers with different intervals
